i feel like dying
I tell myself that it going to get better but it doesnt. It always like this. An infinite loopy mess.
Like a higher thing dindt want me to be good at this.
"You sound lamentable and pathetic. Believe believe believe" is a type of argument that those who at their end of their cage after the
'suffering' can say. not me not me not me. Because I dont know when ill finally be in a confort cage of peace. it doesnt exist  and if it does,
 my mind refuses it. Because that confort leads to ignorance of the evil I hate. The ones that hurt. I wont have the time to protect
  myself. So It refuses. It refuses it. The confort cage, it refuses it.